Add files via upload
Write a method in the MusicExchangeCenter class called songsByPopularity() that returns (i.e., not displays) an ArrayList of Pair<Integer,Song> objects where the key of the pair is an Integer representing the number of downloads and the value is the Song object. The integer key should represent the number of times that the song was downloaded. The list returned should be sorted in decreasing order according to the number of times the song was downloaded. To sort a list of Pair objects, you can use the following code:
Collections.sort(yourListOfPairs, new Comparator<Pair<Integer, Song>>() {
public int compare(Pair<Integer, Song> p1, Pair<Integer, Song> p2) {
// PUT YOUR CODE IN HERE
}
});



Just insert the missing code so that it returns an appropriate integer indicating whether pair p1 comes before or after pair p2 in the sort order (see notes, Chapter 8, Page 281).